refactor(auth): 認証ミドルウェアを resolveAuth + requireAuth の2層構造に分離
- sessionAuth（検証+ガード一体）を resolveAuth（グローバル）+ requireAuth（ガード）に分離
- resolveAuth を全リクエストに適用し、JWT検証を1箇所に集約
- chat.ts と /auth/me の手書きJWT検証を削除（c.get("adminUser") に統一）
- broadcast.ts の型ハック（as never）を解消
- AuthUser の import 元を ~/schemas/auth-schema に統一
- テスト用 withResolveAuth ヘルパーを追加
