+++ server/src/index.ts
@@ -5,7 +5,12 @@ import { handleLineEvent, handleR2Event, handleScheduled } from "~/handlers";
 import type { R2EventMessage } from "~/handlers/r2-event-handler";
 import { logger } from "~/lib/logger";
 import { getSentryOptions } from "~/lib/sentry";
-import { corsMiddleware, errorHandler, securityHeaders } from "~/middleware";
+import {
+  corsMiddleware,
+  errorHandler,
+  resolveAuth,
+  securityHeaders,
+} from "~/middleware";
 import {
   authRoutes,
   broadcastAdminRoutes,
@@ -26,6 +31,7 @@ const app = new OpenAPIHono<{ Bindings: CloudflareBindings }>();
 
 app.use("*", corsMiddleware);
 app.use("*", securityHeaders);
+app.use("*", resolveAuth);
 
 app.onError(errorHandler);
