+++ server/src/middleware/index.ts
@@ -1,3 +1,4 @@
+export { resolveAuth } from "./auth";
 export { corsMiddleware } from "./cors";
 export { errorHandler } from "./error-handler";
 export { lineSignatureVerify } from "./line-signature";
