+++ server/src/routes/admin/knowledge/index.ts
@@ -1,6 +1,6 @@
 import { OpenAPIHono } from "@hono/zod-openapi";
 
-import { sessionAuth } from "~/middleware/session-auth";
+import { requireAuth } from "~/middleware/auth";
 import { knowledgeConvertRoutes } from "./convert";
 import { knowledgeFilesRoutes } from "./files";
 import { knowledgeSyncRoutes } from "./sync";
@@ -9,7 +9,7 @@ export const knowledgeAdminRoutes = new OpenAPIHono<{
   Bindings: CloudflareBindings;
 }>();
 
-knowledgeAdminRoutes.use("*", sessionAuth);
+knowledgeAdminRoutes.use("*", requireAuth);
 
 knowledgeAdminRoutes.route("/", knowledgeSyncRoutes);
 knowledgeAdminRoutes.route("/", knowledgeFilesRoutes);
